third_party/WebKit/LayoutTests/fast/encoding/api/surrogate-pairs.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>

description("Test invalid UTF-16 surrogate pairs with UTF-8 encoding");

var badStrings = [
    { input: "'abc123'", expected: [97, 98, 99, 49, 50, 51] }, // Sanity check.
    { input: "'\\ud800'", expected: [0xef, 0xbf, 0xbd] }, // Surrogate half.
    { input: "'\\udc00'", expected: [0xef, 0xbf, 0xbd] }, // Surrogate half.
    { input: "'abc\\ud800123'", expected: [0x61, 0x62, 0x63, 0xef, 0xbf, 0xbd, 0x31, 0x32, 0x33] }, // Surrogate half.
    { input: "'abc\\udc00123'", expected: [0x61, 0x62, 0x63, 0xef, 0xbf, 0xbd, 0x31, 0x32, 0x33] }, // Surrogate half.
    { input: "'\\udc00\\ud800'",  expected: [239, 191, 189, 239, 191, 189] } // Wrong order.
];

badStrings.forEach(
    function(t) {
        evalAndLog("encoded = new TextEncoder('utf-8').encode(" + t.input + ")");
        shouldBeEqualToString("JSON.stringify([].slice.call(encoded))", JSON.stringify(t.expected));
        debug("");
    });

</script>
<script src="../../js/resources/js-test-post.js"></script>
